create new file for graphql types
